Fix bug introduced in commit c57002f in article-definition
When a new article is added but the code/law is only defined in the reference,
the new method did not more find the code/law. This patch walk to the
(first) reference inside the same verb.

Probably a more robust method should be implemented in the future, possibly
in a visitor with proper rules to find the edited code/law.