Uniformisation of method for new alineas
When new alineas are added there were two slightly different cursors for the
end cursor: -1 or len(old_content), which means a difference of 1 character
(the last newline) => -1 everywhere.

Fix a small bug introduced recently (after fixing alineas inside the text):
alineas added to articles were now added in the beginning of the article =>
fixed this by checking if the reference scope is the whole article (or a
specific part inside the article).

New articles are now treated as new words, similarly to other operations, in
order to benefit to common typography operations.