Exact diffs
Exact diffs are the very words added or removed by the amendments -- not
a diff created by an external independant diff program, which could be not
exactly the very same words of the amendment, e.g. because there are repeated
words and the external diff program balance differently the added/suppressed
words.

Exact diffs are useful for a future git-blame-like feature and they are easier
to merge given they are smaller and character-based, and it can be assumed that
they are collectively compatible (else the amendments would have fallen
“amendement tombé” in the French legal dialect).

Here the diff format used is very similar to the standard GNU diff output
(“Unidiff”), but it is character-based instead of line-based, beginning from
character 1. Hence partly like https://github.com/google/diff-match-patch/wiki/Unidiff
but only with the first rule (and removed/added words are on their standalone
line instead of being on the same line, for improved readability).

Issue: #3