Possibility to apply the diff on an arbitrary text
Currently the AddDiffVisitor required a text supplied by Archeo Lex,
but this is only available for texts in force; if you want to apply
DuraLex and then SedLex on a projet/proposition de loi, you need to
supply the text of the specific article of the projet/proposition de
loi: this is the goal of this amendment to SedLex.

This is implemented quickly with a new parameter in the constructor,
setting the text on which the diff will be applied. On the longer
term other means could be created, for instance another visitor setting
a ppl/pjl similarly to what AddArcheoLexFilenameVisitor currently does.