Encode added or removed articles in the diff header
Previously this was encoded only in the filename header
(--- and +++), now it is also encoded in the text headers
(@@).

With this new feature, the filename headers could be removed
when the context about the filename is already known, as in
Durafront where each diff is encapsulated in a JSON structure
by law name and article name.