Wrong search context when editType == 'add' and alinea-reference
Because of an edge case due to the previous resolution of #5, when
editType == 'add' but we are setting an alinea-ref, the condition
set the next alinea instead of the current.

This resolution of #5 now add new words at the end of the window,
and consequently the 'après les mots' word-ref now set the end
window cursor with the value of the begin cursor. When the edit
node comes, only the end cursor is important when words (and
sentences) are added.

This mostly reverts bf05202 but not entirely, and the alternative
mechanism is implemented.

Issue: #5