Properly encapsulate the amendment in a bill-article, add documentation, remove state variable 'filename'
Until now SedLex was mostly applied on in-force laws, whose the text was
externally retrieved through Archéo Lex, but this was not really amendments
on a bill article. This commit fixes how things should be organised to use
SedLex on an amendment applied on a bill article, mostly the high-level
tree.

Also, in AddDiffVisitor, the state variable 'filename' is removed, because
not really useful and it is a limiting factor to apply some edit operation
on multiple articles (this is still not available, but this commit is a
step forwards).