In-force laws can now be retrieved from the standard dialect of Archéo Lex
The previous dialect of Archéo Lex texts was “one article per file”, it was
easier to manage in a first time, but this dialect is more extensive to
compute in Archéo Lex and is not currently available on https://archeo-lex.fr.

On the other side, the dialect “whole law on a single file” is the standard
dialect of Archéo Lex and is currently updated every day on
https://archeo-lex.fr; it is also not that complicated to parse given it is
Markdown texts. From this commit, SedLex can now read both dialects.

This main dialect is however more extensive from a SedLex point of view given
the whole law/code is read instead of one or some articles.